fix: improve SSR handling by moving PostHog identification to client component
